// src/state-management/Auth/login/login.types.ts
export enum LOGIN_ACTION_TYPES {
  SET_LOGIN_USER = "SET_LOGIN_USER",
  SET_IS_SUBMITING = "SET_IS_SUBMITING",
  SET_ERROR_MESSAGE = "SET_ERROR_MESSAGE",
  POST_LOGIN_START = "POST_LOGIN_START",
  POST_LOGIN_SUCCESS = "POST_LOGIN_SUCCESS",
  POST_LOGIN_FAILED = "POST_LOGIN_FAILED",
  RESET_SIGNIN = "RESET_SIGNIN",
  SET_MODAL_IS_OPEN = "SET_MODAL_IS_OPEN",
  SET_OTP = "SET_OTP",
  SET_OTP_ERROR = "SET_OTP_ERROR",
  SET_EMAIL_VALIDATE = "SET_EMAIL_VALIDATE",
  REQUEST_LOGIN_OTP_START = "REQUEST_LOGIN_OTP_START",
  REQUEST_LOGIN_OTP_SUCCESS = "REQUEST_LOGIN_OTP_SUCCESS",
  REQUEST_LOGIN_OTP_FAILED = "REQUEST_LOGIN_OTP_FAILED",
  RESET_SHOW_MSG = "RESET_SHOW_MSG",
}

export type LoginData = {
  username: "";
  password: "";
};

export type RegisterData = {
  registerData: LoginData;
};

export type EventObj = {
  name: string;
  value: string;
};